minor changes in deleting the brands using modal

+++ myshop/products/models.py
@@ -5,7 +5,7 @@ from datetime import datetime
 class Product(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(50), nullable=False )
-    price = db.Column(db.Numeric(10,2), nullable=False)
+    price = db.Column(db.Integer, nullable=False)
     discount=db.Column(db.Integer, default=0 )
     description = db.Column(db.Text, nullable=False)
     stock = db.Column(db.Integer, nullable=False)
